+++ test/runner.nim
@@ -1,0 +1,6 @@
+## Test harness for Dup
+##
+## Author: Josh Girvin
+## License: MIT
+
+import ./docker_test